add mse loss without loss base class
