+++ neuralop/losses/finite_diff.py
@@ -1,0 +1,60 @@
+import torch
+"""
+finite_diff.py implements utilities for computing derivatives via finite-difference method
+"""
+
+#Set fix{x,y,z}_bnd if function is non-periodic in {x,y,z} direction
+#x: (*, s)
+#y: (*, s)
+def central_diff_1d(x, h, fix_x_bnd=False):
+    dx = (torch.roll(x, -1, dims=-1) - torch.roll(x, 1, dims=-1))/(2.0*h)
+
+    if fix_x_bnd:
+        dx[...,0] = (x[...,1] - x[...,0])/h
+        dx[...,-1] = (x[...,-1] - x[...,-2])/h
+    
+    return dx
+
+#x: (*, s1, s2)
+#y: (*, s1, s2)
+def central_diff_2d(x, h, fix_x_bnd=False, fix_y_bnd=False):
+    if isinstance(h, float):
+        h = [h, h]
+
+    dx = (torch.roll(x, -1, dims=-2) - torch.roll(x, 1, dims=-2))/(2.0*h[0])
+    dy = (torch.roll(x, -1, dims=-1) - torch.roll(x, 1, dims=-1))/(2.0*h[1])
+
+    if fix_x_bnd:
+        dx[...,0,:] = (x[...,1,:] - x[...,0,:])/h[0]
+        dx[...,-1,:] = (x[...,-1,:] - x[...,-2,:])/h[0]
+    
+    if fix_y_bnd:
+        dy[...,:,0] = (x[...,:,1] - x[...,:,0])/h[1]
+        dy[...,:,-1] = (x[...,:,-1] - x[...,:,-2])/h[1]
+        
+    return dx, dy
+
+#x: (*, s1, s2, s3)
+#y: (*, s1, s2, s3)
+def central_diff_3d(x, h, fix_x_bnd=False, fix_y_bnd=False, fix_z_bnd=False):
+    if isinstance(h, float):
+        h = [h, h, h]
+
+    dx = (torch.roll(x, -1, dims=-3) - torch.roll(x, 1, dims=-3))/(2.0*h[0])
+    dy = (torch.roll(x, -1, dims=-2) - torch.roll(x, 1, dims=-2))/(2.0*h[1])
+    dz = (torch.roll(x, -1, dims=-1) - torch.roll(x, 1, dims=-1))/(2.0*h[2])
+
+    if fix_x_bnd:
+        dx[...,0,:,:] = (x[...,1,:,:] - x[...,0,:,:])/h[0]
+        dx[...,-1,:,:] = (x[...,-1,:,:] - x[...,-2,:,:])/h[0]
+    
+    if fix_y_bnd:
+        dy[...,:,0,:] = (x[...,:,1,:] - x[...,:,0,:])/h[1]
+        dy[...,:,-1,:] = (x[...,:,-1,:] - x[...,:,-2,:])/h[1]
+    
+    if fix_z_bnd:
+        dz[...,:,:,0] = (x[...,:,:,1] - x[...,:,:,0])/h[2]
+        dz[...,:,:,-1] = (x[...,:,:,-1] - x[...,:,:,-2])/h[2]
+        
+    return dx, dy, dz
+
